Renamed components, cleanup
