+++ frontend/components/lines/LinesDetailMetrics/styles.module.css
@@ -1,4 +1,23 @@
 /* * */
+/* HEADING SECTION */
+
+.headingSection {
+	display: flex;
+	flex-direction: column;
+	justify-content: flex-start;
+	align-items: flex-start;
+	gap: var(--size-spacing-15);
+}
+
+.headingSectionRow {
+	display: flex;
+	flex-direction: row;
+	justify-content: flex-start;
+	align-items: center;
+	gap: var(--size-spacing-15);
+}
+
+/* * */
 /* CONTAINER */
 
 .container {
