+++ frontend/components/lines/LinesDetailPath/styles.module.css
