+++ frontend/components/lines/LinesDetailPathList/index.tsx
