+++ frontend/translations/pt-PT.json
@@ -106,6 +106,9 @@
 				"subtitle": "executabilidade do serviço de todos os percursos desta linha nos últimos 15 dias"
 			}
 		},
+		"LinesDetailPathMap": {
+			"active_vehicles_counter": "{count, plural, =0 {Nenhum veículo em circulação} one {# veículo em circulação} other {# veículos em circulação}}"
+		},
 		"SelectActivePatternGroup":{
 			"placeholder": "Selecionar percurso/destino..."
 		},
@@ -114,9 +117,6 @@
 		},
 		"AlertsCarousel": {
 			"heading": "Alertas de Serviço"
-		},
-		"LineMap": {
-			"active_vehicles_counter": "{count, plural, =0 {Nenhum veículo em circulação} one {# veículo em circulação} other {# veículos em circulação}}"
 		},
 		"expand_filters": "Expandir Filtros",
 		"collapse_filters": "Esconder Filtros"
